Initial commit
Mostly planning and a few bits of code for PoC. Basic framework for CLI
in `linker`, test code for reading nether portals from a minecraft world
in `search_region.py`, some initial planning for input config setup in
`portals.conf` (this needs to be better defined in future), and the
start of an object-oriented structure for the backend in `PortalMap.py`.

Next steps are to flesh out more of the back end code, and get
documentation in as I go with some docstrings.