Moved to RTree implementation for PortalMap
Changed from QuadTree to RTree implementation, will require libspatialindex as a
dependency for the project. Implemented first part of link check, getting a list
of all portals within a linkable range. The second part is simply finding the
closest portal by euclidean distance; still unsure of the best implementation
for this -- it would be nice to use the RTree nearest neighbour search for
efficiency, but if using the entire RTree there is no guarantee that the nearest
neighbour is in range, so some extra logic is needed. It may be simpler to
manually calculate the distances for all portals passing the range check but
there will be instances where this needs to process a very large number of
portals so may be inefficient. Perhaps the answer is to choose the simple
implementation and optimise later if required.